[entity-browser4] don't browse into objects without suggestions
Current browse rules are strikingly similar to infer-block-type. If we
add another block type that handles another type we need to teach the
entity browser to do that in 2 places. Probably the Block types should
be registered and they should decide whether they handle this data or
not. Then we only need 1 action, add the block type.