Router4-based apps can now mount on any pages with a given URL prefix
If an Electric app was designed to mount on `/` then `r/router` used to match
all routes against `/` only. Router4 now respects `document.baseURI` and the
`<base>` element. The same app can now mount under `/foo/` given `<base
href="/foo/">` is present in the document's <head>. `r/link`, `r/Navigate!` and
`r/ReplaceState!` all account for the document's base.

This feature is somewhat similar to React-Router's `basename`, defining a common
prefix for all routes of a router. But unlike React-Router, `r/router` will
still attempt to match routes even if the document is served from a URL not
matching the document's `baseURI`, acknowledging that a document might be served
from a different URL than the one used to resolve relative links. I.e.:

- With `<base href="/foo/">`
- On `http://localhost/bar` – really "/bar", not matching "/foo/bar"
- r/router will ignore the "misalignment" and r/route will be ['bar]
- React-Router will not match, strictly expecting `/foo/bar`.

Under this advanced use case, it is up to the user to canonicalize links or
perform a redirect.