[forms] fix user validation message not displaying on invalid inputs
Validation was set on the input as a `data-errormessage` attribute. The user was
supposed to display it using CSS (e.g. with `content: attr()`). But <input> does
not support pseudoelements so there was no pure CSS way to reach and display the
input's data-errormessage. Now we set the data-errormessage onto the <input>'s
parent iff there can be no conflict with another input, otherwise we leave it
onto the input and users will have to render validation error themselves.