[entity-browser] nav as late as possible
There's 3 stages, if used optimally they do the following
- run query -> return entity IDs
- nav -> return a fat object we can pull
- pull -> turn the object into a data map

Fat object might not always be a thing, or might be `select *` for SQL
etc. There's differences with data sources.

This commit pushes nav to later stages so we only nav on visible data.
This won't improve performance on datomic much as the navs are
`d/entity` calls which are cheap, but could be significant on e.g.
sakila SQL demo where the nav is a `select *` today.