[navigator] can now sort .getTag column on vars
Make sort more resilient by downcasting common non-comparable types to a common
comparable representation (e.g. to string). This is extensible from userland.
API to be defined.

`.getTag` is a property of clojure.lang.Var. It usually return nil or a java
class, but can also be a string (e.g. "[[Ljava.lang.Object;") or an ident (e.g.
for isa? and hierarchies).